Hub: add outcome converters
