+++ packages/hub/src/wallet/db/utils.ts
@@ -1,5 +1,3 @@
-import {isAllocationOutcome, Outcome} from '@statechannels/nitro-protocol';
-import {AssetOutcome} from '@statechannels/nitro-protocol/lib/src/contract/outcome';
 import * as Knex from 'knex';
 
 export function addAddressCheck(knex: Knex, table: string, column: string) {
@@ -14,18 +12,4 @@ export function addBytesCheck(knex: Knex, table: string, column: string) {
     ALTER TABLE ${table}\
     ADD CONSTRAINT ${column}_is_bytes CHECK (${column} ~ '^0x[0-9a-fA-f]*$')
   `);
-}
-
-function assetOutcomeAddPriorities(assetOutcome: AssetOutcome) {
-  if (isAllocationOutcome(assetOutcome)) {
-    const allocationsWithPriorities = assetOutcome.allocation.map((allocationItem, index) => ({
-      ...allocationItem,
-      priority: index
-    }));
-    return {...assetOutcome, allocation: allocationsWithPriorities};
-  }
-}
-
-export function outcomeAddPriorities(outcome: Outcome) {
-  return outcome.map(assetOutcomeAddPriorities);
 }
